Prts 1677 aggregator reseller migration fixes (#31815)
* PRTS-1677 aggregator to reseller migration fixes

* PRTS-1677 adding function descriptions

* PRTS-1677 adding metrics for success and failure cases

* PRTS-1677 adding experimentation

* PRTS-1677 resolving comments

* PRTS-1677 updating failed test case and metric dimension

* Update app/Models/Merchant/Service.php


* Update app/Models/Merchant/Service.php


* Updating test cases and comments

* adding merchant users data sync validation

* PRTS-1677 resolving minor comments

* PRTS-1677 removing unwanted imports

* PRTS-1677 changing variable name to partner

* PRTS-1677 fix test case exception

* PRTS-1677 adding metrics for success and failure cases

* PRTS-1677 adding experimentation

* PRTS-1677 resolving comments

* Update app/Models/Merchant/Service.php


* Update app/Models/Merchant/Service.php


* adding merchant users data sync validation

* PRTS-1677 resolving minor comments

* PRTS-1677 removing unwanted imports

* Revert "PRTS-1677 adding metrics for success and failure cases"

This reverts commit d6adbeeb6563a901ccb32aaba22d3a6c57d5c184.

* PRTS-1677 merging master

* PRTS-1677 master merge file name update

* PRTS-1677 fix test case

* PRTS-1677 updated to batch events for bulk endpoint

* PRTS-1677 updating service layer

* PRTS-1677 adding input validation

* PRTS-1677 UPDATING TEST CASES

+++ app/Services/AuthService.php
@@ -127,7 +127,7 @@ class AuthService
         return $this->sendRequest('applications', Requests::GET, $input);
     }
 
-    public function deleteApplication(string $id, string $merchantId) : array
+    public function deleteApplication(string $id, string $merchantId, bool $deletePartnerMappings = true) : array
     {
         $input = [Application\Entity::MERCHANT_ID => $merchantId];
 
@@ -139,13 +139,14 @@ class AuthService
             ]
         );
         $result = $this->sendRequest('applications/' . $id, Requests::PUT, $input);
+        if ($deletePartnerMappings === true)
+        {
+            // deletes the access mapping for the application
+            (new AccessMap\Core)->deleteAccessMapByApplicationId($id);
 
-        // deletes the access mapping for the application
-        (new AccessMap\Core)->deleteAccessMapByApplicationId($id);
-
-        // delete merchant and application mapping
-        (new MerchantApplications\Core)->deleteByApplication($id);
-
+            // delete merchant and application mapping
+            (new MerchantApplications\Core)->deleteByApplication($id);
+        }
         return $result;
     }
